+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 ### Added
+- Progress bar `#39`
+- `Tools/Continuous Avatar Uploader` `#40`
 
 ### Changed